Complete archives.go rewrite
Use the correct archiver according to the mimeType.
Handle almost (besides brotli) archive formats.
Don't try to print contents of a single file compressed (not a tar
archive).
Differentiate `.tar.gz` and `.gz` according to regex matching of the
file's extension.
Use a wsearch inspired algorithm to print in equal widths the table of
files inside the archives.